+++ flutter/digit-ui-components/digit_components/lib/widgets/atoms/labelled_fields.dart
@@ -1,5 +1,6 @@
 import 'package:digit_ui_components/digit_components.dart';
 import 'package:digit_ui_components/theme/digit_extended_theme.dart';
+import 'package:digit_ui_components/widgets/atoms/tooltip_2.dart';
 import 'package:flutter/material.dart';
 import '../../utils/utils.dart';
 
@@ -11,9 +12,9 @@ class LabeledField extends StatelessWidget {
   final TextStyle? labelStyle;
   final TextStyle? descriptionStyle;
   final EdgeInsets? padding;
-  final bool preferToolTipBelow;
+  final TooltipPosition tooltipPosition;
   final String? infoText;
-  final TooltipTriggerMode tooltipTriggerMode;
+  final TooltipTrigger tooltipTrigger;
   final bool wrapLabelText;
   final MainAxisAlignment mainAxisAlignment;
   final CrossAxisAlignment crossAxisAlignment;
@@ -31,8 +32,8 @@ class LabeledField extends StatelessWidget {
     this.padding,
     this.isRequired = false,
     this.infoText,
-    this.preferToolTipBelow = false,
-    this.tooltipTriggerMode = TooltipTriggerMode.tap,
+    this.tooltipPosition = TooltipPosition.bottomCenter,
+    this.tooltipTrigger = TooltipTrigger.onTap,
     this.wrapLabelText = true,
     this.mainAxisAlignment = MainAxisAlignment.start,
     this.crossAxisAlignment = CrossAxisAlignment.start,
@@ -59,6 +60,7 @@ class LabeledField extends StatelessWidget {
         children: [
           if (label != null)
             Wrap(
+              crossAxisAlignment: WrapCrossAlignment.center,
               children: [
                 RichText(
                   text: TextSpan(
@@ -85,13 +87,17 @@ class LabeledField extends StatelessWidget {
                 ),
                 if (infoText != null) const SizedBox(width: spacer1),
                 if (infoText != null)
-                  Tooltip(
-                    message: infoText,
-                    preferBelow: preferToolTipBelow,
-                    triggerMode: tooltipTriggerMode,
-                    child: const Icon(
-                      Icons.info_outline,
-                      size: spacer4,
+                  SizedBox(
+                    height: 24,
+                    width: 24,
+                    child: DigitTooltip(
+                      tooltipPosition: tooltipPosition,
+                      trigger: tooltipTrigger,
+                      contentDescription: infoText,
+                      child: const Icon(
+                        Icons.info_outline,
+                        size: spacer4,
+                      ),
                     ),
                   ),
               ],
@@ -142,14 +148,17 @@ class LabeledField extends StatelessWidget {
                     ),
                   if (infoText != null) const SizedBox(width: spacer1),
                   if (infoText != null)
-                    Tooltip(
-                      message: infoText,
-                      preferBelow: preferToolTipBelow,
-                      triggerMode: tooltipTriggerMode,
-                      child: Icon(
-                        Icons.info_outline,
-                        size: spacer5,
-                        color: theme.colorTheme.text.secondary,
+                    SizedBox(
+                      height: 24,
+                      width: 24,
+                      child: DigitTooltip(
+                        tooltipPosition: tooltipPosition,
+                        trigger: tooltipTrigger,
+                        contentDescription: infoText,
+                        child: const Icon(
+                          Icons.info_outline,
+                          size: spacer4,
+                        ),
                       ),
                     ),
                 ],